Refactor Seller module namespaces and folder structure
Standardized namespaces and reorganized Seller Dtos, Handlers, and Queries into feature-based subfolders (complaint, dashboard, order, product, revenue). Fixed incorrect namespace references and enabled BuyerAssemblyMarker registration in MediatR. Also removed unused folder entries from the project file.
