+++ DPTS.Applications/Buyer/BuyerAssemblyMarker.cs
@@ -1,0 +1,6 @@
+﻿namespace DPTS.Applications.Buyer
+{
+    public sealed class BuyerAssemblyMarker
+    {
+    }
+}
