--- DPTS.Applications/Seller/Dtos/ComplaintOverviewDto.cs
@@ -1,11 +1,0 @@
-﻿namespace DPTS.Applications.Seller.conplaints.Dtos
-{
-    public class ComplaintOverviewDto
-    {
-       public string Name { get; set; } = string.Empty;
-        public decimal Value { get; set; }
-        public string Information { get; set; } = string.Empty;
-    }
-
-}
-
