--- DPTS.Applications/Seller/Dtos/RevenueChartPointDto.cs
@@ -1,9 +1,0 @@
-﻿namespace DPTS.Applications.Sellers.revenues.Dtos
-{
-    public class RevenueChartPointDto
-    {
-        public DateTime Date { get; set; }
-        public decimal Revenue { get; set; }
-    }
-
-}
