--- DPTS.Applications/Seller/Dtos/RevenueOverviewDto .cs
@@ -1,10 +1,0 @@
-namespace DPTS.Applications.Sellers.revenues.Dtos
-{
-    public class RevenueOverviewDto
-    {
-        public string NameOverview { get; set; } = string.Empty;
-        public decimal Values { get; set; }
-        public string Information { get; set; } = string.Empty;
-    }
-
-}
