--- DPTS.Applications/Seller/Dtos/SellerOverviewDto.cs
@@ -1,9 +1,0 @@
-﻿namespace DPTS.Applications.Sellers.overviews.Dtos
-{
-    public class SellerOverviewDto
-    {
-        public string OverViewName { get; set; } = string.Empty;
-        public string Information { get; set; } = string.Empty;
-        public decimal Value { get; set; }
-    }
-}
