+++ DPTS.Applications/Seller/Dtos/complaint/ComplaintOverviewDto.cs
@@ -1,0 +1,11 @@
+﻿namespace DPTS.Applications.Seller.Dtos.complaint
+{
+    public class ComplaintOverviewDto
+    {
+       public string Name { get; set; } = string.Empty;
+        public decimal Value { get; set; }
+        public string Information { get; set; } = string.Empty;
+    }
+
+}
+
