+++ DPTS.Applications/Seller/Dtos/dashboard/SellerOverviewDto.cs
@@ -1,0 +1,9 @@
+﻿namespace DPTS.Applications.Seller.Dtos.dashboard
+{
+    public class SellerOverviewDto
+    {
+        public string OverViewName { get; set; } = string.Empty;
+        public string Information { get; set; } = string.Empty;
+        public decimal Value { get; set; }
+    }
+}
