+++ DPTS.Applications/Seller/Dtos/order/OrderListItemDto.cs
@@ -1,0 +1,12 @@
+﻿namespace DPTS.Applications.Seller.Dtos.order
+{
+    public class OrderListItemDto
+    {
+        public string EscrowId { get; set; } = string.Empty;
+        public string BuyerName { get; set; } = string.Empty;
+        public decimal Amount { get; set; }
+        public string Status { get; set; } = string.Empty;
+        public string TimeBuyAt { get; set; } = string.Empty;
+        public string DateBuyAt { get; set; } = string.Empty;
+    }
+}
