+++ DPTS.Applications/Seller/Dtos/revenue/RevenueChartPointDto.cs
@@ -1,0 +1,9 @@
+﻿namespace DPTS.Applications.Seller.Dtos.revenue
+{
+    public class RevenueChartPointDto
+    {
+        public DateTime Date { get; set; }
+        public decimal Revenue { get; set; }
+    }
+
+}
