+++ DPTS.Applications/Seller/Dtos/revenue/RevenueOverviewDto .cs
@@ -1,0 +1,10 @@
+namespace DPTS.Applications.Seller.Dtos.revenue
+{
+    public class RevenueOverviewDto
+    {
+        public string NameOverview { get; set; } = string.Empty;
+        public decimal Values { get; set; }
+        public string Information { get; set; } = string.Empty;
+    }
+
+}
