--- DPTS.Applications/Seller/Query/SearchSellerOrdersQuery.cs
@@ -1,6 +1,0 @@
-﻿namespace DPTS.Applications.Seller.orders.Queries
-{
-    public class SearchSellerOrdersQuery
-    {
-    }
-}
