+++ DPTS.Applications/Seller/Query/product/GetTopSellingProductsQuery.cs
@@ -1,0 +1,11 @@
+﻿using DPTS.Applications.Seller.Dtos.product;
+using DPTS.Applications.Shareds;
+using MediatR;
+
+namespace DPTS.Applications.Seller.Query.product
+{
+    public class GetTopSellingProductsQuery : IRequest<ServiceResult<IEnumerable<TopSellingProductDto>>>
+    {
+        public string SellerId { get; set; } = string.Empty;
+    }
+}
