+++ DPTS.Applications/Seller/Query/product/SearchSellerOrdersQuery.cs
@@ -1,0 +1,6 @@
+﻿namespace DPTS.Applications.Seller.Query.product
+{
+    public class SearchSellerOrdersQuery
+    {
+    }
+}
