+++ DPTS.Applications/Seller/Query/revenue/GetRevenueOverviewQuery.cs
@@ -1,0 +1,11 @@
+﻿using DPTS.Applications.Seller.Dtos.revenue;
+using DPTS.Applications.Shareds;
+using MediatR;
+
+namespace DPTS.Applications.Seller.Query.revenue
+{
+    public class GetRevenueOverviewQuery: IRequest<ServiceResult<IEnumerable<RevenueOverviewDto>>>
+    {
+        public string SellerId { get; set; } = string.Empty;
+    }
+}
